refactor(require/module): further separation of require and module
